+++ packages/editor/web/src/RichContentEditor/blockStyleFn.js
@@ -18,6 +18,8 @@ const types = {
   'header-six': 'headerSix',
   atomic: 'atomic',
   'code-block': 'codeBlock',
+  'ordered-list-item': 'orderedList',
+  'unordered-list-item': 'unorderedList',
 };
 const isList = type => {
   return type === 'ordered-list-item' || type === 'unordered-list-item';
@@ -42,24 +44,17 @@ export default (theme, styleToClass) => {
       data: { textAlignment, dynamicStyles = {} },
     } = contentBlock.toJS();
 
-    let classList;
+    const key = types[type] || 'text';
+    const classList = [styles[key], theme[key]];
 
-    if (isList(type)) {
-      classList = [
+    if (type !== 'atomic') {
+      classList.push(
         styles[textAlignment],
         theme[textAlignment],
-        listAlignmentClass(textAlignment, getTextDirection(text)),
-      ];
-    } else {
-      const key = types[type] || 'text';
-      classList = [styles[key], theme[key]];
-
-      if (type !== 'atomic') {
-        classList.push(styles[textAlignment], theme[textAlignment], [
-          depthClassName(depth),
-          textBlockAlignmentClass(textAlignment, getTextDirection(text)),
-        ]);
-      }
+        isList(type)
+          ? listAlignmentClass(textAlignment, getTextDirection(text))
+          : [depthClassName(depth), textBlockAlignmentClass(textAlignment, getTextDirection(text))]
+      );
     }
 
     const dynamicClasses = Object.entries(dynamicStyles).map(styleToClass);